apps/woob-extractor/requirements.txt
boto3~=1.24.75
jmespath~=1.0.1
PyYAML~=6.0
woob~=3.0
pdfminer.six==20220524
pyhumps~=3.8.0
requests~=2.28.1